[fix] 정규식 제거
