+++ FrontEnd/my-car/src/mycar/routers/Color.js
@@ -33,7 +33,7 @@ function Color() {
   const dispatch = useUserCarAction();
   const userCar = useUserCarState();
 
-  const [loading, setLoading] = useState(true);
+  const [loading, setLoading] = useState(false);
   const action = useContext(ColorActionContext);
   useEffect(() => {
     if (!userCar.outerColor?.name && !userCar.innerColor?.name) {
@@ -51,34 +51,6 @@ function Color() {
         select: inColorOptions[0],
       });
     }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  const CarColor = ['abyss', 'silver', 'blue', 'brown', 'gray', 'white'];
-
-  const preloadCarImg = () => {
-    const promise = CarColor.map((color) => {
-      return new Promise((resolve) => {
-        const imagePaths = Array.from(
-          { length: 60 },
-          (_, index) =>
-            `https://s3.ap-northeast-2.amazonaws.com/hyundaimycar.store/rotation/${color}/${
-              index + 1
-            }.webp`,
-        );
-        imagePaths.forEach((path) => {
-          const img = new Image();
-          img.src = path;
-          img.onload = resolve;
-        });
-      });
-    });
-    return Promise.all(promise);
-  };
-  useLayoutEffect(() => {
-    preloadCarImg().then(() => {
-      setLoading(false);
-    });
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
