+++ FrontEnd/my-car/src/mycar/routers/Engine.js
@@ -6,6 +6,7 @@ import { Container } from './Model';
 import { MYCAR } from '../../constant';
 import TrimImg from '../components/common/TrimImg';
 import { useSelect } from '../hook/useSelect';
+import { useLayoutEffect } from 'react';
 
 const EngineContainer = styled(Container)`
   flex-direction: row;
@@ -44,6 +45,34 @@ function Engine() {
     field: 'engine',
     page,
   });
+
+  const CarColor = ['abyss', 'silver', 'blue', 'brown', 'gray', 'white'];
+
+  const preloadCarImg = () => {
+    const promise = CarColor.map((color) => {
+      return new Promise((resolve) => {
+        const imagePaths = Array.from(
+          { length: 60 },
+          (_, index) =>
+            `https://s3.ap-northeast-2.amazonaws.com/hyundaimycar.store/rotation/${color}/${
+              index + 1
+            }.webp`,
+        );
+        imagePaths.forEach((path) => {
+          const img = new Image();
+          img.src = path;
+          img.onload = resolve;
+        });
+      });
+    });
+    return Promise.all(promise);
+  };
+  useLayoutEffect(() => {
+    preloadCarImg().then(() => {
+      // setLoading(false);
+    });
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
   return (
     <EngineContainer>
